Write Puppet provisioners
So... I turned my Vagrantfile to spaghetti. Because some versions of
Vagrant (specifically the one I'm using, 1.8.4) don't play nice with
Puppet 4.x.x (it tries to use a deprecated switch...), I wrote a
disgusting method out of anger to just do it on the commandline
of the guest. Don't worry, I have plans to make all of this go away...